LayoutTests/inspector/audit/basic-promise-expected.txt

Testing the basic functionality of audits.


== Running test suite: Audit.Basic
-- Running test case: Audit.Basic.Promise.Boolean.True
Testing value `new Promise((resolve, reject) => resolve(true))`...
PASS: Result should be "pass".

-- Running test case: Audit.Basic.Promise.String.Pass
Testing value `new Promise((resolve, reject) => resolve("pass"))`...
PASS: Result should be "pass".

-- Running test case: Audit.Basic.Promise.Object.Pass
Testing value `new Promise((resolve, reject) => resolve({level: "pass"}))`...
PASS: Result should be "pass".
